+++ CHANGELOG.md
@@ -7,6 +7,10 @@
 * New data version check, will automatically run the exp fix command if the playerdata is older. 
 
 **Changed:**
+* Command to add ability and shotlocks now require a boolean for permanency, these will prevail after a level reset or an experience fix.
+  * /kingdomkeys ability give <ability> <permanent> [player]
+  * /kingdomkeys shotlock give <shotlock> <permanent> [player]
+  * The command to take them away will remove them both temporary and permanent.
 * Flowmotion renamed to Wall Kick, equip more to allow more bounces.
 * Flowmotion-related abilities into growth abilities.
 * Wall kick obtained at levels 7 and 51 for Warrior, 9 and 62 for Mystic, 14 and 43 for Guardian.